requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
asttokens==3.0.1
azure-core==1.39.0
azure-cosmos==4.15.0
azure-functions==1.24.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.23.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
decorator==5.2.1
distro==1.9.0
Django==5.2.13
executing==2.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
groq==1.2.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipython==9.10.1
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jiter==0.14.0
lxml==6.1.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
msal==1.36.0
msal-extensions==1.3.1
numpy==2.4.4
openai==2.32.0
pandas==3.0.2
parso==0.8.6
pexpect==4.9.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
RapidFuzz==3.14.5
requests==2.33.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
tenacity==9.1.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.8
